Remove default directories. Set default vhosts.
We don't want to keep /var/www/html or /var/www/default on our base
docker image. This removes those directories.

We do want our custom vhost conf file to be available in the apache
sites-available directory and set as a symlink in the sites-enabled
directory so we allow puppet to manage that file and move it into place.

We also want to set environment variables for Drupal if we are serving
over SSL. This is set via a file in apache's conf.d directory.