Add composer path to docker image PATH.
We want the docker image to have the composer bin directory available in
the path when commands are run by users. To do this, we set our
pakcer-built docker container to be an intermediary container and then
use a docker file to set the environment variable and save a new
container intended for use.